feat: migrate taglines to its own directory, general cleanup of widget, and fix isssue with taglines not showing on anonymous accounts
